Collapse use-archive-agent.pure.ts into use-archive-agent.ts
